Refactor Identity with DbContext and perform code cleanup

+++ server/Social/src/Social.Application/Interfaces/ICacheKeyService.cs
@@ -2,6 +2,6 @@
 {
     public interface ICacheKeyService
     {
-        string GenerateFriendStatusCacheKey(int userId1, int userId2);
+        string GenerateFriendStatusCacheKey(long userId1, long userId2);
     }
 }
